feat: add tests for decoded and direct message capture routing and subscription
