+++ tests/packet_capture/test_core_behavior.py
@@ -12,11 +12,6 @@ from meshcore_packet_capture.enums import PayloadType
 from meshcore_packet_capture.packet_capture import PacketCapture
 
 
-@pytest.fixture
-def capture() -> PacketCapture:
-    return PacketCapture(enable_mqtt=False)
-
-
 def test_retry_delay_uses_backoff_and_jitter(monkeypatch: pytest.MonkeyPatch, capture: PacketCapture) -> None:
     capture.connection_retry_delay = 2
     capture.connection_retry_backoff_multiplier = 2.0
@@ -83,10 +78,6 @@ def test_resolve_topic_template_replaces_token_placeholder(monkeypatch: pytest.M
 
 def test_get_topic_command_requires_explicit_config(capture: PacketCapture) -> None:
     assert capture.get_topic("command", broker_num=1) is None
-
-
-def test_get_topic_decoded_requires_explicit_config(capture: PacketCapture) -> None:
-    assert capture.get_topic("decoded", broker_num=1) is None
 
 
 @pytest.mark.asyncio
@@ -175,388 +166,6 @@ async def test_publish_status_includes_packet_stats_aliases(capture: PacketCaptu
     assert published
     assert published[0]["stats"]["packets_received"] == 1234
     assert published[0]["stats"]["packets_sent"] == 567
-
-
-@pytest.mark.asyncio
-async def test_handle_decoded_message_event_publishes_decoded_payload(
-    capture: PacketCapture,
-) -> None:
-    published: list[tuple[str | None, str, object | None, int | None]] = []
-
-    capture.enable_mqtt = True
-    capture.mqtt_connected = True
-    capture.device_name = "node"
-    capture.device_public_key = "abc123"
-    client_obj = object()
-    capture.mqtt_clients = [{"client": client_obj, "broker_num": 1, "label": "mqtt1"}]
-
-    def _get_topic(topic_type, broker_num=None):
-        assert broker_num == 1
-        if topic_type == "decoded":
-            return "meshcore/private/ABC123/decoded"
-        if topic_type == "direct":
-            return "meshcore/private/ABC123/direct"
-        if topic_type == "channel":
-            return "meshcore/private/ABC123/channel/{CHANNEL}"
-        raise AssertionError(f"unexpected topic_type: {topic_type}")
-
-    def _publish(topic, payload, **kwargs):
-        published.append((topic, payload, kwargs.get("client"), kwargs.get("broker_num")))
-        return {"attempted": 1, "succeeded": 1}
-
-    capture.get_topic = _get_topic  # type: ignore[method-assign]
-    capture.safe_publish = _publish
-
-    event = types.SimpleNamespace(
-        type="CONTACT_MSG_RECV",
-        payload={
-            "type": "PRIV",
-            "text": "hello world",
-            "from": "node-a",
-            "pubkey_prefix": "a1b2c3",
-            "msg_id": "msg-1",
-        },
-    )
-
-    await capture.handle_decoded_message_event(event)
-
-    assert published
-    assert published[0][0] == "meshcore/private/ABC123/decoded"
-    assert published[1][0] == "meshcore/private/ABC123/direct"
-    assert published[0][2] is client_obj
-    assert published[0][3] == 1
-    payload_json = json.loads(published[0][1])
-    assert payload_json["type"] == "DECODED_MESSAGE"
-    assert payload_json["direction"] == "direct"
-    assert payload_json["message"] == "hello world"
-    assert payload_json["from"] == "node-a"
-
-
-@pytest.mark.asyncio
-async def test_handle_decoded_message_event_includes_payload_signal(
-    capture: PacketCapture,
-) -> None:
-    published: list[tuple[str | None, str]] = []
-
-    capture.enable_mqtt = True
-    capture.mqtt_connected = True
-    capture.mqtt_clients = [{"client": object(), "broker_num": 1, "label": "mqtt1"}]
-
-    def _get_topic(topic_type, broker_num=None):
-        if topic_type == "direct":
-            return "meshcore/private/ABC123/direct"
-        if topic_type == "channel":
-            return "meshcore/private/ABC123/channel/{CHANNEL}"
-        return None
-
-    def _publish(topic, payload, **kwargs):
-        published.append((topic, payload))
-        return {"attempted": 1, "succeeded": 1}
-
-    capture.get_topic = _get_topic  # type: ignore[method-assign]
-    capture.safe_publish = _publish
-
-    event = types.SimpleNamespace(
-        type="CONTACT_MSG_RECV",
-        payload={
-            "type": "PRIV",
-            "text": "hello world",
-            "from": "node-a",
-            "pubkey_prefix": "a1b2c3",
-            "msg_id": "msg-1",
-            "snr": 12.5,
-            "rssi": -87,
-        },
-    )
-
-    await capture.handle_decoded_message_event(event)
-
-    payload_json = json.loads(published[0][1])
-    assert payload_json["snr"] == 12.5
-    assert payload_json["rssi"] == -87.0
-
-
-@pytest.mark.asyncio
-async def test_handle_decoded_message_event_does_not_use_rf_cache_signal(
-    capture: PacketCapture,
-) -> None:
-    published: list[tuple[str | None, str]] = []
-
-    capture.enable_mqtt = True
-    capture.mqtt_connected = True
-    capture.mqtt_clients = [{"client": object(), "broker_num": 1, "label": "mqtt1"}]
-    capture.rf_data_cache = {
-        "old": {"snr": 1.0, "rssi": -100, "timestamp": 1.0},
-        "recent": {"snr": 7.25, "rssi": -72, "timestamp": 10.0},
-    }
-
-    def _get_topic(topic_type, broker_num=None):
-        if topic_type == "direct":
-            return "meshcore/private/ABC123/direct"
-        if topic_type == "channel":
-            return "meshcore/private/ABC123/channel/{CHANNEL}"
-        return None
-
-    def _publish(topic, payload, **kwargs):
-        published.append((topic, payload))
-        return {"attempted": 1, "succeeded": 1}
-
-    capture.get_topic = _get_topic  # type: ignore[method-assign]
-    capture.safe_publish = _publish
-
-    event = types.SimpleNamespace(
-        type="CONTACT_MSG_RECV",
-        payload={
-            "type": "PRIV",
-            "text": "hello world",
-            "from": "node-a",
-            "pubkey_prefix": "a1b2c3",
-            "msg_id": "msg-1",
-        },
-    )
-
-    original_time = pc_mod.time.time
-    try:
-        pc_mod.time.time = lambda: 12.0
-        await capture.handle_decoded_message_event(event)
-    finally:
-        pc_mod.time.time = original_time
-
-    payload_json = json.loads(published[0][1])
-    assert "snr" not in payload_json
-    assert "rssi" not in payload_json
-
-
-@pytest.mark.asyncio
-async def test_handle_decoded_message_event_reads_metadata_signal(
-    capture: PacketCapture,
-) -> None:
-    published: list[tuple[str | None, str]] = []
-
-    capture.enable_mqtt = True
-    capture.mqtt_connected = True
-    capture.mqtt_clients = [{"client": object(), "broker_num": 1, "label": "mqtt1"}]
-
-    def _get_topic(topic_type, broker_num=None):
-        if topic_type == "direct":
-            return "meshcore/private/ABC123/direct"
-        if topic_type == "channel":
-            return "meshcore/private/ABC123/channel/{CHANNEL}"
-        return None
-
-    def _publish(topic, payload, **kwargs):
-        published.append((topic, payload))
-        return {"attempted": 1, "succeeded": 1}
-
-    capture.get_topic = _get_topic  # type: ignore[method-assign]
-    capture.safe_publish = _publish
-
-    event = types.SimpleNamespace(
-        type="CONTACT_MSG_RECV",
-        payload={
-            "type": "PRIV",
-            "text": "hello world",
-            "from": "node-a",
-            "pubkey_prefix": "a1b2c3",
-            "msg_id": "msg-1",
-            "metadata": {
-                "signal": {
-                    "SNR": 3.14,
-                    "RSSI": -91,
-                }
-            },
-        },
-    )
-
-    await capture.handle_decoded_message_event(event)
-
-    payload_json = json.loads(published[0][1])
-    assert payload_json["snr"] == 3.14
-    assert payload_json["rssi"] == -91.0
-
-
-@pytest.mark.asyncio
-async def test_handle_decoded_message_event_routes_direct_topic_per_broker(
-    capture: PacketCapture,
-) -> None:
-    published: list[tuple[str | None, str, object | None, int | None]] = []
-
-    capture.enable_mqtt = True
-    capture.mqtt_connected = True
-    client_obj = object()
-    capture.mqtt_clients = [{"client": client_obj, "broker_num": 1, "label": "mqtt1"}]
-
-    def _get_topic(topic_type, broker_num=None):
-        assert broker_num == 1
-        if topic_type == "decoded":
-            return "meshcore/private/ABC123/decoded"
-        if topic_type == "direct":
-            return "meshcore/private/ABC123/direct"
-        if topic_type == "channel":
-            return "meshcore/private/ABC123/channel/{CHANNEL}"
-        return None
-
-    def _publish(topic, payload, **kwargs):
-        published.append((topic, payload, kwargs.get("client"), kwargs.get("broker_num")))
-        return {"attempted": 1, "succeeded": 1}
-
-    capture.get_topic = _get_topic  # type: ignore[method-assign]
-    capture.safe_publish = _publish  # type: ignore[method-assign]
-
-    event = types.SimpleNamespace(
-        type="CONTACT_MSG_RECV",
-        payload={
-            "type": "PRIV",
-            "text": "hello world",
-            "from": "node-a",
-            "pubkey_prefix": "a1b2c3",
-            "msg_id": "msg-1",
-        },
-    )
-
-    await capture.handle_decoded_message_event(event)
-
-    assert published
-    assert published[0][0] == "meshcore/private/ABC123/decoded"
-    assert published[1][0] == "meshcore/private/ABC123/direct"
-    assert published[0][2] is client_obj
-    assert published[0][3] == 1
-
-
-@pytest.mark.asyncio
-async def test_handle_decoded_message_event_routes_channel_topic_per_broker(
-    capture: PacketCapture,
-) -> None:
-    published: list[tuple[str | None, str]] = []
-
-    capture.enable_mqtt = True
-    capture.mqtt_connected = True
-    capture.mqtt_clients = [{"client": object(), "broker_num": 1, "label": "mqtt1"}]
-
-    def _get_topic(topic_type, broker_num=None):
-        assert broker_num == 1
-        if topic_type == "decoded":
-            return "meshcore/private/ABC123/decoded"
-        if topic_type == "direct":
-            return "meshcore/private/ABC123/direct"
-        if topic_type == "channel":
-            return "meshcore/private/ABC123/channel/{CHANNEL}"
-        return None
-
-    def _publish(topic, payload, **_kwargs):
-        published.append((topic, payload))
-        return {"attempted": 1, "succeeded": 1}
-
-    capture.get_topic = _get_topic  # type: ignore[method-assign]
-    capture.safe_publish = _publish  # type: ignore[method-assign]
-
-    event = types.SimpleNamespace(
-        type="CHANNEL_MSG_RECV",
-        payload={
-            "type": "CHAN",
-            "text": "hello channel",
-            "from": "node-a",
-            "channel_idx": 3,
-            "msg_id": "msg-2",
-        },
-    )
-
-    await capture.handle_decoded_message_event(event)
-
-    assert published
-    assert published[0][0] == "meshcore/private/ABC123/decoded"
-    assert published[1][0] == "meshcore/private/ABC123/channel/3"
-
-
-@pytest.mark.asyncio
-async def test_setup_event_handlers_does_not_subscribe_message_events(
-    monkeypatch: pytest.MonkeyPatch, capture: PacketCapture
-) -> None:
-    subscribed: list[tuple[str, object]] = []
-
-    event_type = types.SimpleNamespace(
-        RX_LOG_DATA="RX_LOG_DATA",
-        RAW_DATA="RAW_DATA",
-        STATUS_RESPONSE="STATUS_RESPONSE",
-        CONTACT_MSG_RECV="CONTACT_MSG_RECV",
-        CHANNEL_MSG_RECV="CHANNEL_MSG_RECV",
-        DISCONNECTED="DISCONNECTED",
-    )
-    monkeypatch.setattr(pc_mod, "EventType", event_type)
-
-    capture.meshcore = types.SimpleNamespace(
-        subscribe=lambda event_name, handler: subscribed.append((event_name, handler)),
-        dispatcher=types.SimpleNamespace(subscriptions=[]),
-        unsubscribe=lambda _subscription: None,
-    )
-
-    await capture.setup_event_handlers()
-
-    subscribed_names = {name for name, _handler in subscribed}
-    assert "CONTACT_MSG_RECV" not in subscribed_names
-    assert "CHANNEL_MSG_RECV" not in subscribed_names
-
-
-@pytest.mark.asyncio
-async def test_setup_event_handlers_subscribes_contact_when_direct_configured(
-    monkeypatch: pytest.MonkeyPatch, capture: PacketCapture
-) -> None:
-    subscribed: list[tuple[str, object]] = []
-
-    event_type = types.SimpleNamespace(
-        RX_LOG_DATA="RX_LOG_DATA",
-        RAW_DATA="RAW_DATA",
-        STATUS_RESPONSE="STATUS_RESPONSE",
-        CONTACT_MSG_RECV="CONTACT_MSG_RECV",
-        CHANNEL_MSG_RECV="CHANNEL_MSG_RECV",
-        DISCONNECTED="DISCONNECTED",
-    )
-    monkeypatch.setattr(pc_mod, "EventType", event_type)
-    monkeypatch.setenv("PACKETCAPTURE_MQTT1_ENABLED", "true")
-    monkeypatch.setenv("PACKETCAPTURE_MQTT1_TOPIC_DIRECT", "meshcore/private/{PUBLIC_KEY}/direct")
-
-    capture.meshcore = types.SimpleNamespace(
-        subscribe=lambda event_name, handler: subscribed.append((event_name, handler)),
-        dispatcher=types.SimpleNamespace(subscriptions=[]),
-        unsubscribe=lambda _subscription: None,
-    )
-
-    await capture.setup_event_handlers()
-
-    subscribed_names = {name for name, _handler in subscribed}
-    assert "CONTACT_MSG_RECV" in subscribed_names
-    assert "CHANNEL_MSG_RECV" not in subscribed_names
-
-
-@pytest.mark.asyncio
-async def test_setup_event_handlers_subscribes_channel_when_channel_configured(
-    monkeypatch: pytest.MonkeyPatch, capture: PacketCapture
-) -> None:
-    subscribed: list[tuple[str, object]] = []
-
-    event_type = types.SimpleNamespace(
-        RX_LOG_DATA="RX_LOG_DATA",
-        RAW_DATA="RAW_DATA",
-        STATUS_RESPONSE="STATUS_RESPONSE",
-        CONTACT_MSG_RECV="CONTACT_MSG_RECV",
-        CHANNEL_MSG_RECV="CHANNEL_MSG_RECV",
-        DISCONNECTED="DISCONNECTED",
-    )
-    monkeypatch.setattr(pc_mod, "EventType", event_type)
-    monkeypatch.setenv("PACKETCAPTURE_MQTT1_ENABLED", "true")
-    monkeypatch.setenv("PACKETCAPTURE_MQTT1_TOPIC_CHANNEL", "meshcore/private/{PUBLIC_KEY}/channel/{CHANNEL}")
-
-    capture.meshcore = types.SimpleNamespace(
-        subscribe=lambda event_name, handler: subscribed.append((event_name, handler)),
-        dispatcher=types.SimpleNamespace(subscriptions=[]),
-        unsubscribe=lambda _subscription: None,
-    )
-
-    await capture.setup_event_handlers()
-
-    subscribed_names = {name for name, _handler in subscribed}
-    assert "CONTACT_MSG_RECV" not in subscribed_names
-    assert "CHANNEL_MSG_RECV" in subscribed_names
 
 
 def test_on_mqtt_connect_subscribes_command_topic(
